removing unnecessary files and line of codes
