+++ lib/model/pressure_data_model.dart
@@ -1,0 +1,8 @@
+class PressureData {
+  final int date;
+  final List<int> value;
+  PressureData({
+    required this.date,
+    required this.value,
+  });
+}
