+++ lib/services/state_provider.dart
@@ -1,37 +1,34 @@
 import 'package:flutter_blue_plus/flutter_blue_plus.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:pressureruuvi/functions/csv_generator.dart';
+import 'package:pressureruuvi/model/pressure_data_model.dart';
 import 'package:stop_watch_timer/stop_watch_timer.dart';
 
+// Return the map of Sensor Containers which collects data
 final clickedIconProvider = StateProvider<Map<String, bool>>((ref) {
   return {};
 });
 
+// Controls the StopWatchTimer package for each sensor
 final stopWatchTimerProvider =
     StateProvider.family<StopWatchTimer, String>((ref, deviceName) {
   return StopWatchTimer();
 });
 
+// List of connected sensors
 final devicesProvider = StateProvider<List<BluetoothDevice>>((ref) {
   return [];
 });
 
+// List of sensors which are currently gathering data
 final devicesListProvider = StateProvider<List<String>>((ref) {
   return [];
 });
 
-final isRefreshingProvider = StateProvider<bool>((ref) => false);
-
+// Holds the data of each sensor
 final devicesDataProvider =
     StateProvider<Map<String, List<List<PressureData>>>>((ref) {
   return {};
 });
 
-final streamFromDevices =
-    StreamProvider<Map<String, List<List<PressureData>>>>((ref) {
-  final state = ref.watch(devicesDataProvider);
-
-  return Stream.value(state);
-});
-
+// Check if the process is ongoing
 final isLoadingProvider = StateProvider<bool>((ref) => false);
